Improved solver with faster processing
Updated the game solver to be faster by using multiprocessing for checking each part of the game board at the same time and threading for quicker decision-making on moves. Also made some tweaks to how the mouse moves and checks if the game board has settled. This version scored 2x better than the previous version.