improved accuracy and match logic
tweaked the detection of the game board to handle different screen sizes and improved the accuracy of object matching by adjusting the template matching threshold. these changes help in better identification of objects on the board, leading to more effective gameplay. tried to add solver implementation of chain reactions and future moves.